+++ src/newio_rethinkdb/__init__.py
@@ -2,6 +2,6 @@ from rethinkdb import *  # noqa
 from .net import Connection, ConnectionPool  # noqa
 
 
-def set_loop_type_curio():
+def set_loop_type():
     import rethinkdb.net
     rethinkdb.net.connection_type = Connection