+++ tasks.py
@@ -2,31 +2,17 @@ from invoke import task
 
 
 @task
-def install(ctx):
-    ctx.run('pip install -r requirements.txt')
-    ctx.run('pre-commit install')
+def test(ctx):
+    ctx.run('pytest --cov=newio_rethinkdb')
 
 
 @task
-def lint(ctx):
-    ctx.run('pre-commit run --all-files')
+def build(ctx):
+    ctx.run('rm -f dist/*')
+    ctx.run('python setup.py sdist')
 
 
 @task
-def test(ctx, cov=False, verbose=False):
-    cov = ' --cov=curethinkdb --cov-report=term-missing' if cov else ''
-    verbose = ' -v -x --log-level=debug' if verbose else ''
-    cmd = f'pytest --tb=short{cov}{verbose} tests'
-    ctx.run(cmd)
-
-
-@task
-def dist(ctx, upload=False):
-    cmds = [
-        'rm -f dist/*',
-        'python setup.py bdist_wheel',
-    ]
-    if upload:
-        cmds.append('twine upload dist/*')
-    for cmd in cmds:
-        ctx.run(cmd, echo=True)
+def publish(ctx):
+    build(ctx)
+    ctx.run('twine upload dist/*')