+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2018-07-01
+### Added
+- Copy and modified from CuRethinkDB
+- Add basic tests
+
 ## [0.0.1] - 2017-12-25
 ### Added
 - Open Source on GitHub